Publish basic status of file writer `Master`

+++ src/KafkaW.h
@@ -178,7 +178,7 @@ public:
   ProducerTopic(ProducerTopic &&);
   ProducerTopic(std::shared_ptr<Producer> producer, std::string name);
   ~ProducerTopic();
-  int produce(uchar *msg_data, int msg_size, bool print_err = false);
+  int produce(uchar *msg_data, size_t msg_size, bool print_err = false);
   int produce(std::unique_ptr<Producer::Msg> &msg);
   // Currently it's nice to have access to these for statistics:
   std::shared_ptr<Producer> producer;
